+++ tests/test_config.py
@@ -1,7 +1,7 @@
 """
 Unit tests for sat.config
 
-(C) Copyright 2019-2020 Hewlett Packard Enterprise Development LP.
+(C) Copyright 2019-2021 Hewlett Packard Enterprise Development LP.
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -29,9 +29,20 @@ import unittest
 from unittest import mock
 
 import sat
-from sat.config import ConfigValidationError, DEFAULT_CONFIG_PATH, get_config_value, read_config_value_file, \
-    generate_default_config, load_config, SATConfig, SAT_CONFIG_SPEC, validate_log_level, _option_value, \
-    OptionSpec
+from sat.config import (
+    ConfigFileExistsError,
+    ConfigValidationError,
+    DEFAULT_CONFIG_PATH,
+    OptionSpec,
+    SATConfig,
+    SAT_CONFIG_SPEC,
+    _option_value,
+    generate_default_config,
+    get_config_value,
+    load_config,
+    read_config_value_file,
+    validate_log_level
+)
 from tests.common import ExtendedTestCase
 
 CONFIGS_DIR = os.path.join(os.path.dirname(__file__), 'resources/configs')
@@ -129,14 +140,13 @@ class TestGenerateDefaultConfig(unittest.TestCase):
         mock.patch('sat.config.SAT_CONFIG_SPEC', self.fake_config_spec).start()
         self.mock_open = mock.patch('builtins.open').start()
         self.mock_output_stream = self.mock_open.return_value
-        self.mock_getenv = mock.patch('sat.config.os.getenv', return_value=DEFAULT_CONFIG_PATH).start()
         self.mock_isdir = mock.patch('sat.config.os.path.isdir', return_value=True).start()
         self.mock_isfile = mock.patch('sat.config.os.path.isfile', return_value=False).start()
         self.mock_makedirs = mock.patch('sat.config.os.makedirs').start()
 
         self.expected_config = dedent("""\
         # Default configuration file for SAT.
-        # (C) Copyright 2019-2020 Hewlett Packard Enterprise Development LP.
+        # (C) Copyright 2019-2021 Hewlett Packard Enterprise Development LP.
 
         # Permission is hereby granted, free of charge, to any person obtaining a
         # copy of this software and associated documentation files (the "Software"),
@@ -167,26 +177,22 @@ class TestGenerateDefaultConfig(unittest.TestCase):
 
     def test_basic_config_generation(self):
         """Test basic default config generation"""
-        generate_default_config()
+        generate_default_config(DEFAULT_CONFIG_PATH)
         self.mock_open.assert_called_with(DEFAULT_CONFIG_PATH, 'w')
         self.mock_output_stream.write.assert_called_once_with(self.expected_config)
         self.mock_makedirs.assert_not_called()
 
     def test_generate_alternate_location(self):
-        """Test generate_default_config() can write to the file specified by SAT_CONFIG_FILE"""
-        self.mock_getenv.return_value = '/tmp/sat.toml'
-        generate_default_config()
-        self.mock_getenv.assert_called_with('SAT_CONFIG_FILE', DEFAULT_CONFIG_PATH)
-        self.mock_open.assert_called_once_with('/tmp/sat.toml', 'w')
+        """Test generate_default_config() can write to any given path"""
+        generate_default_config('/mock/path.toml')
+        self.mock_open.assert_called_once_with('/mock/path.toml', 'w')
         self.mock_output_stream.write.assert_called_once_with(self.expected_config)
         self.mock_makedirs.assert_not_called()
 
     def test_generate_create_directory(self):
         """Test generate_default_config() will create a config directory if needed"""
-        self.mock_getenv.return_value = '/etc/opt/cray/sat.toml'
         self.mock_isdir.return_value = False
-        generate_default_config()
-        self.mock_getenv.assert_called_with('SAT_CONFIG_FILE', DEFAULT_CONFIG_PATH)
+        generate_default_config('/etc/opt/cray/sat.toml')
         self.mock_open.assert_called_once_with('/etc/opt/cray/sat.toml', 'w')
         self.mock_output_stream.write.assert_called_once_with(self.expected_config)
         self.mock_makedirs.assert_called_once_with('/etc/opt/cray', exist_ok=True)
@@ -194,15 +200,19 @@ class TestGenerateDefaultConfig(unittest.TestCase):
     def test_generate_with_username(self):
         """Test generating config with a username will write a config file with a username"""
         self.expected_config = self.expected_config.replace('# username = ""', 'username = "sat_user"')
-        generate_default_config(username='sat_user')
+        generate_default_config(DEFAULT_CONFIG_PATH, username='sat_user')
         self.mock_open.assert_called_with(DEFAULT_CONFIG_PATH, 'w')
         self.mock_output_stream.write.assert_called_once_with(self.expected_config)
         self.mock_makedirs.assert_not_called()
 
     def test_generate_file_exists(self):
-        """Test generating a config file when the file exists will not overwrite"""
+        """Test generating a config file when the file exists will not overwrite and raises ConfigFileExistsError."""
         self.mock_isfile.return_value = True
-        generate_default_config()
+        with self.assertRaisesRegex(
+                ConfigFileExistsError,
+                f'Configuration file "{DEFAULT_CONFIG_PATH}" already exists. Not generating configuration file.'
+        ):
+            generate_default_config(DEFAULT_CONFIG_PATH)
         self.mock_open.assert_not_called()
         self.mock_output_stream.write.assert_not_called()
         self.mock_makedirs.assert_not_called()
@@ -210,7 +220,7 @@ class TestGenerateDefaultConfig(unittest.TestCase):
     def test_generate_file_exists_force(self):
         """Test generating a config file when the file exists will overwrite if forcing"""
         self.mock_isfile.return_value = True
-        generate_default_config(force=True)
+        generate_default_config(DEFAULT_CONFIG_PATH, force=True)
         self.mock_open.assert_called_with(DEFAULT_CONFIG_PATH, 'w')
         self.mock_output_stream.write.assert_called_once_with(self.expected_config)
         self.mock_makedirs.assert_not_called()